+++ core/components/romanesco/elements/plugins/07_computations/c_content/manipulatedom.plugin.php
@@ -127,11 +127,14 @@ switch ($modx->event->name) {
         // Remove rows from grids that have a reversed column order on mobile
         $dom->filter('.ui.reversed.grid > .row')->unwrapInner();
 
-        // If grids are stackable on tablet, also hide designated mobile elements
+        // If grids are stackable on tablet, also hide (or show) designated mobile elements
         $dom->filter('.ui[class*="stackable on tablet"].grid [class*="mobile hidden"]')
             ->removeClass('mobile')
             ->removeClass('hidden')
             ->addClass('tablet or lower hidden')
+        ;
+        $dom->filter('.ui[class*="stackable on tablet"].grid [class*="mobile only"]')
+            ->addClass('tablet only')
         ;
 
         // If grids are stackable on tablet, some responsive image sizes might be incorrect